requirements.txt
pandas
streamlit
numpy
google-cloud-bigquery
openpyxl
xlrd
python-dotenv
db-dtypes
xlsxwriter
plotly
